requirements.txt
black==22.6.0
typer==0.6.1
imageio==2.19.3
pydantic==1.9.0
matplotlib==3.5.2
ipython==8.6.0
notebook==6.5.2
pandas==1.4.4
albumentations==1.1.0
torchtyping==0.1.4
wandb>=0.13.5